Tests added

+++ src/main/java/com/ecommerce/year2_sem2_project/Factory/ProductFactory.java
@@ -3,7 +3,7 @@ package com.ecommerce.year2_sem2_project.Factory;
 import com.ecommerce.year2_sem2_project.Entity.Product;
 
 public class ProductFactory {
-    public static Product createProduct(String name, String description, Long price, int stock, String imageUrl) {
+    public static Product createProduct(String name, String description, double price, int stock, String imageUrl) {
         Product product = new Product();
         product.setName(name);
         product.setDescription(description);
